requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# AI/ML for image analysis
torch>=2.0.0
torchvision>=0.15.0

# GitHub API
PyGithub>=2.1.0

# Scheduling
schedule>=1.2.0

# Logging
colorlog>=6.7.0
